Connected SQL.
Changed product.cs attributes due to values on SQL table. Parameterless constructor to assist SQL data integration.

Styling and clean code is next step. PRIO2

+++ E-Comemrce/Utility.cs
@@ -70,7 +70,7 @@ namespace E_Commerce
                     if (!string.IsNullOrEmpty(input))
                     {
 
-                        number = Convert.ToInt32(input);
+                        number = Convert.ToDouble(input);
                         return number;
 
                     }
